backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pytz==2023.3
tzdata==2023.3
# 认证相关依赖
PyJWT==2.8.0
email-validator==2.1.0
phonenumbers==8.13.26
requests==2.31.0
cryptography==41.0.7

# 系统监控依赖
psutil==5.9.6

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1 